+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Show a readable reason when signing in fails, such as "Incorrect username or password", instead of internal error details
+
 ## [0.0.2] - 2026-09-14
 
 ### Added